packages/python-rag/requirements.txt
langchain
langchain-community
langchain-mistralai
langchain-openai
faiss-cpu
tiktoken
openai
mistralai
python-dotenv
tenacity